fnc_arcor_md_customer/requirements.txt
functions-framework==3.*
google-cloud>=0.34.0
google-cloud-storage>=2.7.0
google-cloud-bigquery>=2.7.0